+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All relevant changes to `mateusjunges/laravel-kafka` will be documented here.
 
+## [2023-02-09 v1.12.0](https://github.com/mateusjunges/laravel-kafka/compare/v1.11.0...v1.12.0)
+- Allow consumers to subscribe to `SIGQUIT` and `SIGTERM` signals by @mateusjunges on [#172](https://github.com/mateusjunges/laravel-kafka/pull/172)
+- Add `onStopConsume` method to the Consumer class
+
 ## [2023-01-30 v1.11.0](https://github.com/mateusjunges/laravel-kafka/compare/v1.10.2...v1.11.0)
 - Add support for Laravel 10 by @mateusjunges and @mihaileu on [#171](https://github.com/mateusjunges/laravel-kafka/pull/171)